improve large tree performance in defaultEquality
This should fix #10 and fix #18, and should address performance issues on larger trees. On my test tree of probably 3000+ nodes, `defaultEquality` and `angular.copy` inside of it were the biggest loads on the CPU in the profiler. 

Note that angular.copy in 1.2 was shallow, and in 1.3 was changed to be deep.

https://docs.angularjs.org/api/ng/function/angular.copy
https://docs.angularjs.org/guide/migration
https://github.com/angular/angular.js/commit/b59b04f98a0b59eead53f6a53391ce1bbcbe9b57

This was especially problematic since we were be making deep copies of the entire object and then throwing away the childrenNodes just to compare the parent - for a top level folder with many nested subfolders, we were doing a lot of work for little gain :(

Courtesy of stidges: https://github.com/stidges/angular-tree-control/blob/23bafaad7a58b76d581f1a02eda48e34e8100b98/angular-tree-control.js#L49